Create success.php
Create success.php that user is redirected to after successful account creation in signup.php

Add if statement in signup.php to confirm if password and password confirmation inputs are a match

+++ public/signup.php
@@ -3,34 +3,42 @@ require_once("../templates/header.php");
 ?>
 
 <?php
+$passErr="";
 //IF STATEMENT
 //Check if POST data is set, and if so execute below code block
 if(isset($_POST['create'])) {
-    //require sanitize function
-    require_once ("../src/sanitize.php");
-    //TRY STATEMENT
-    try {
-        //require DBConnect
-        require_once("../src/DBConnect.php");
-        //Create $newUser array and pass values into keys from POST superglobal
-        //Data will first be passed as an argument into the sanitize function, cleaned and then returned to array
-        $newUser = array (
-            "username" => sanitize($_POST['username']),
-            "email" => sanitize($_POST['email']),
-            "address" => sanitize($_POST['address']),
-            "password" => sanitize($_POST['password'])
-        );
+    if($_POST['password'] == $_POST['passwordConfirm']) {
+        //require sanitize function
+        require_once ("../src/sanitize.php");
+        //TRY STATEMENT
+        try {
+            //require DBConnect
+            require_once("../src/DBConnect.php");
+            //Create $newUser array and pass values into keys from POST superglobal
+            //Data will first be passed as an argument into the sanitize function, cleaned and then returned to array
+            $newUser = array (
+                "username" => sanitize($_POST['username']),
+                "email" => sanitize($_POST['email']),
+                "address" => sanitize($_POST['address']),
+                "password" => sanitize($_POST['password'])
+            );
 
-        //Insert values from $new_user array into database as a prepared SQL statement
-        $sql = sprintf( "INSERT INTO %s (%s) values (%s)", "users", 
-               implode(", ",array_keys($newUser)), ":" . implode(", :", array_keys($newUser)) );
-        $statement = $connection->prepare($sql);
-        $statement->execute($newUser);
-        
-    } catch (PDOException $error) {
-        echo $sql . "<br>" . $error->getMessage();
+            //Insert values from $new_user array into database as a prepared SQL statement
+            $sql = sprintf( "INSERT INTO %s (%s) values (%s)", "users", 
+                implode(", ",array_keys($newUser)), ":" . implode(", :", array_keys($newUser)) );
+            $statement = $connection->prepare($sql);
+            $statement->execute($newUser);
+
+            header("location:success.php");
+            exit;
+            
+        } catch (PDOException $error) {
+            echo $sql . "<br>" . $error->getMessage();
+        }
+    } else {
+        $passErr= ("Passwords not do match. Please try again.");
     }
-}
+} 
 ?>
 <title>Signup page</title>
 </head>
@@ -53,23 +61,23 @@ if(isset($_POST['create'])) {
                 <!--SIGNUP FORM-->
                 <form action="" method="POST" id="signup-form">
                     <div class="form-floating">
-                        <input name ="username" type="text" class="form-control" id="floatingUserName" required>
+                        <input name ="username" type="text" class="form-control" id="floatingUserName" placeholder="" required>
                         <label for="floatingUserName">Username</label>
                     </div>
                     <div class="form-floating mt-3">
-                        <input name="email" type="email" class="form-control" id="floatingEmail" required>
+                        <input name="email" type="email" class="form-control" id="floatingEmail" placeholder="" required>
                         <label for="floatingEmail">Email</label>
                     </div>
                     <div class="form-floating mt-3">
-                        <input name="address" type="text" class="form-control" id="floatingAddress" required>
+                        <input name="address" type="text" class="form-control" id="floatingAddress" placeholder="" required>
                         <label for="floatingAddress">Address</label>
                     </div>
                     <div class="form-floating mt-3">
-                        <input name="password" type="password" class="form-control" id="floatingPassword" required>
+                        <input name="password" type="password" class="form-control" id="floatingPassword" placeholder="" required>
                         <label for="floatingUPassword">Password</label>
                     </div>
                     <div class="form-floating mt-3">
-                        <input type="password" class="form-control" id="floatingPasswordConfirm" required>
+                        <input name="passwordConfirm" type="password" class="form-control" id="floatingPasswordConfirm" placeholder="" required>
                         <label for="floatingPasswordConfirm">Confirm password</label>
                     </div>
                 </form>
@@ -85,6 +93,7 @@ if(isset($_POST['create'])) {
                 <button type="reset" form="signup-form" class="btn btn-secondary ms-2">Reset</button>
             </div>
         </div>
+        <h5 class="text-danger mt-4"><?php echo($passErr);?></h5>
     </div>
 <?php
     require_once("../templates/footer.php");
